// src/app/models/user.ts
export interface User {
  id: string;
  nom: string;
  prenom: string;
  email: string;
  adresse: string;
  ville: string;
  date_naissance: Date;
  etablissement: String;
  imageUrl:String;
  niveau_scolaire: String;
  telephone: String;
  module:String,
  codePostal:String,
  bio:String,
  balance:number
}